refactored routes such that we now have different files for different functionalities
